// core/RUNTIME/TEMPORAL/types.ts
export interface DurationData {
  years?: number;
  months?: number;
  days?: number;
  hours?: number;
  minutes?: number;
  seconds?: number;
  milliseconds?: number;
}